backend/requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
greenlet==1.1.3
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.2.0
pip==21.2.4
pyparsing==3.0.9
pytesseract==0.3.10
setuptools==58.1.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
zipp==3.11.0
Flask-Mail==0.9.1
Flask-Cors==3.0.10
Flask-Migrate==4.0.0
mariadb==1.1.5.post3
Flask-JWT==0.3.2

celery~=5.2.7
alembic~=1.8.1
PyJWT~=1.4.2